discogs service is refactored
