+++ src/routes/discogs.routes.ts
@@ -1,8 +1,13 @@
 import { Router } from 'express'
-import controller from '../controllers/discogs.controller'
+import DiscogsRepositoryContract from '../repositories/DiscogsRepository'
+import DiscogsService from '../services/DiscogsService'
+import DiscogsController from '../controllers/DiscogsController'
 
+const discogsRepository = new DiscogsRepositoryContract()
+const discogsService = new DiscogsService(discogsRepository)
+const discogsController = new DiscogsController(discogsService)
 const router = Router()
 
-router.post('/', controller.getList)
+router.post('/', discogsController.getDiscogsData)
 
 export default router
